Upload fit data for children rather than parent
