+++ tests/model/test_integrands.py
@@ -1,4 +1,3 @@
-import numpy as np
 import pandas as pd
 
 from cascade.testing_utilities import make_execution_context
@@ -7,7 +6,7 @@ from cascade.dismod.constants import IntegrandEnum
 from cascade.model.integrands import make_average_integrand_cases_from_gbd
 
 
-def test_make_average_integrand_cases_from_gbd(mocker):
+def test_make_average_integrand_cases_from_gbd(mocker, mock_locations):
     get_age_groups = mocker.patch("cascade.model.integrands.get_age_groups")
     get_years = mocker.patch("cascade.model.integrands.get_years")
 
@@ -16,11 +15,12 @@ def test_make_average_integrand_cases_from_gbd(mocker):
     )
     get_years.return_value = [1990, 1995, 2000]
 
-    ec = make_execution_context(parent_location_id=180)
+    ec = make_execution_context(parent_location_id=2, gbd_round_id=6)
 
     average_integrand_cases = make_average_integrand_cases_from_gbd(ec, [1, 2])
 
-    assert np.all(average_integrand_cases.node_id == 180)
+    assert set(average_integrand_cases.node_id) == {5, 6}
+
     for (age_lower, age_upper) in {(0, 1), (1, 4), (4, 82)}:
         for (time_lower, time_upper) in {(1990, 1990), (1995, 1995), (2000, 2000)}:
             for integrand in IntegrandEnum:
@@ -34,7 +34,7 @@ def test_make_average_integrand_cases_from_gbd(mocker):
                 )
 
 
-def test_make_average_integrand_cases_from_gbd__with_birth_prevalence(mocker):
+def test_make_average_integrand_cases_from_gbd__with_birth_prevalence(mocker, mock_locations):
     get_age_groups = mocker.patch("cascade.model.integrands.get_age_groups")
     get_years = mocker.patch("cascade.model.integrands.get_years")
 
@@ -43,7 +43,7 @@ def test_make_average_integrand_cases_from_gbd__with_birth_prevalence(mocker):
     )
     get_years.return_value = [1990, 1995, 2000]
 
-    ec = make_execution_context(parent_location_id=180)
+    ec = make_execution_context(parent_location_id=2, gbd_round_id=6)
 
     average_integrand_cases = make_average_integrand_cases_from_gbd(ec, [1, 2], include_birth_prevalence=True)
 
@@ -51,7 +51,6 @@ def test_make_average_integrand_cases_from_gbd__with_birth_prevalence(mocker):
     assert all(birth_rows.integrand_name == "prevalence")
     assert len(birth_rows) == 3 * 2
 
-    assert np.all(average_integrand_cases.node_id == 180)
     for (age_lower, age_upper) in {(0, 1), (1, 4), (4, 82)}:
         for (time_lower, time_upper) in {(1990, 1990), (1995, 1995), (2000, 2000)}:
             for integrand in IntegrandEnum:
